# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "TravelContext API",
    "summary": "API for the intelligent travel planning DSS prototype.",
    "description": "TravelContext API powers a prototype decision support system for personalized travel planning. The current API version exposes health checks, cities, points of interest, saved profiles, trip requests, personalized recommendations, persisted routes, and decision logs.",
    "contact": {
      "name": "TravelContext project"
    },
    "license": {
      "name": "Academic prototype"
    },
    "version": "0.1.0"
  },
  "servers": [
    {
      "url": "http://127.0.0.1:8000",
      "description": "Local development server"
    }
  ],
  "paths": {
...
    "/trip-requests": {
      "post": {
        "tags": [
          "planning"
        ],
        "summary": "Create trip planning request",
        "description": "Creates a persisted trip planning request from city, reusable user profile, trip duration, and resource constraints. The generated id can be used to request recommendations and routes.",
        "operationId": "create_trip_request_endpoint_trip_requests_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TripRequestCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created trip request.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TripRequestResponse"
                }
              }
            }
          },
          "404": {
            "description": "City was not found.",
            "content": {
              "application/json": {
                "example": {
                  "detail": "City not found"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/recommendations/generate": {
      "post": {
        "tags": [
          "planning"
        ],
        "summary": "Generate personalized recommendations",
        "description": "Generates a ranked list of points of interest for an existing trip request. The scoring model combines interest match, budget fit, route convenience, pace fit, popularity, and data quality, then adjusts the ranking with goals, must-have/avoid constraints, trust level, accessibility needs, and preferred time windows from the saved profile.",
        "operationId": "generate_recommendations_endpoint_recommendations_generate_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RecommendationGenerateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Ranked recommendations with explanations.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecommendationListResponse"
                }
              }
            }
          },
          "404": {
            "description": "Trip request or city was not found.",
            "content": {
              "application/json": {
                "examples": {
                  "tripNotFound": {
                    "summary": "Trip request not found",
                    "value": {
                      "detail": "Trip request not found"
                    }
                  },
                  "cityNotFound": {
                    "summary": "City not found",
                    "value": {
                      "detail": "City not found"
                    }
                  }
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/feedback": {
      "post": {
        "tags": [
          "feedback"
        ],
        "summary": "Save route feedback",
        "description": "Stores a user evaluation for a previously built route. The feedback entry is also written into the persistent decision log together with route snapshots.",
        "operationId": "create_feedback_endpoint_feedback_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FeedbackCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Saved feedback entry.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FeedbackResponse"
                }
              }
            }
          },
          "404": {
            "description": "Route was not found.",
            "content": {
              "application/json": {
                "example": {
                  "detail": "Route not found"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/decision-logs": {
      "get": {
        "tags": [
          "feedback"
        ],
        "summary": "List decision log entries",
        "description": "Returns recent feedback and decision log entries persisted during planning. This is useful for experiments and demonstration of explainable planning history. In the current MVP this endpoint should be treated as a demo/admin surface.",
        "operationId": "list_decision_logs_endpoint_decision_logs_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "description": "Maximum number of log entries to return.",
              "default": 20,
              "title": "Limit"
            },
            "description": "Maximum number of log entries to return."
          }
        ],
        "responses": {
          "200": {
            "description": "Recent decision log entries.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DecisionLogResponse"
                  },
                  "title": "Response List Decision Logs Endpoint Decision Logs Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "DecisionLogResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id",
            "description": "Decision log id.",
            "examples": [
              1
            ]
          },
          "event_type": {
            "type": "string",
            "title": "Event Type",
            "description": "Type of planning event.",
            "examples": [
              "route_built"
            ]
          },
          "city_id": {
            "type": "integer",
            "title": "City Id",
            "description": "City id.",
            "examples": [
              6
            ]
          },
          "trip_request_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Trip Request Id",
            "description": "Related trip request id."
          },
          "recommendation_run_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Recommendation Run Id",
            "description": "Related recommendation run id."
          },
          "route_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Route Id",
            "description": "Related route id."
          },
          "feedback_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Feedback Id",
            "description": "Related feedback id."
          },
          "input_context": {
            "additionalProperties": true,
            "type": "object",
            "title": "Input Context",
            "description": "Request payload and context captured for the event."
          },
          "profile_snapshot": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Profile Snapshot",
            "description": "Expanded profile snapshot used in the decision."
          },
          "recommendations_snapshot": {
            "anyOf": [
              {
                "items": {
                  "additionalProperties": true,
                  "type": "object"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Recommendations Snapshot",
            "description": "Saved recommendation list snapshot if available."
          },
          "route_snapshot": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Route Snapshot",
            "description": "Saved route snapshot if available."
          },
          "explanation_snapshot": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Explanation Snapshot",
            "description": "Explanation fragments captured for auditing."
          },
          "feedback_snapshot": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Feedback Snapshot",
            "description": "Saved feedback snapshot if available."
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At",
            "description": "Decision log creation timestamp."
          }
        },
        "type": "object",
        "required": [
          "id",
          "event_type",
          "city_id",
          "created_at"
        ],
        "title": "DecisionLogResponse"
      },
      "FeedbackCreate": {
        "properties": {
          "route_id": {
            "type": "integer",
            "title": "Route Id",
            "description": "Built route id to evaluate.",
            "examples": [
              1
            ]
          },
          "rating": {
            "type": "integer",
            "maximum": 5.0,
            "minimum": 1.0,
            "title": "Rating",
            "description": "User rating of route usefulness from 1 to 5.",
            "examples": [
              4
            ]
          },
          "comment": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 1000
              },
              {
                "type": "null"
              }
            ],
            "title": "Comment",
            "description": "Optional user comment about the route quality.",
            "examples": [
              "Маршрут хороший, но хотелось бы меньше музеев и больше прогулок."
            ]
          }
        },
        "type": "object",
        "required": [
          "route_id",
          "rating"
        ],
        "title": "FeedbackCreate"
      },
      "FeedbackResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id",
            "description": "Saved feedback id.",
            "examples": [
              1
            ]
          },
          "route_id": {
            "type": "integer",
            "title": "Route Id",
            "description": "Route id.",
            "examples": [
              1
            ]
          },
          "trip_request_id": {
            "type": "integer",
            "title": "Trip Request Id",
            "description": "Trip request id.",
            "examples": [
              1
            ]
          },
          "city_id": {
            "type": "integer",
            "title": "City Id",
            "description": "City id.",
            "examples": [
              6
            ]
          },
          "rating": {
            "type": "integer",
            "maximum": 5.0,
            "minimum": 1.0,
            "title": "Rating",
            "description": "User rating from 1 to 5.",
            "examples": [
              4
            ]
          },
          "comment": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Comment",
            "description": "Optional user comment.",
            "examples": [
              "Маршрут хороший, но хотелось бы больше еды и меньше музеев."
            ]
          },
          "route_points_count": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Route Points Count",
            "description": "Number of points in the evaluated route.",
            "examples": [
              5
            ]
          },
          "estimated_budget": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Estimated Budget",
            "description": "Estimated budget of the evaluated route in rubles.",
            "examples": [
              2950
            ]
          },
          "total_time_minutes": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Total Time Minutes",
            "description": "Total estimated route time in minutes.",
            "examples": [
              481
            ]
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At",
            "description": "Feedback creation timestamp."
          }
        },
        "type": "object",
        "required": [
          "id",
          "route_id",
          "trip_request_id",
          "city_id",
          "rating",
          "route_points_count",
          "estimated_budget",
          "total_time_minutes",
          "created_at"
        ],
        "title": "FeedbackResponse"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "RecommendationGenerateRequest": {
        "properties": {
          "trip_request_id": {
            "type": "integer",
            "title": "Trip Request Id",
            "description": "Previously created trip request id.",
            "examples": [
              1
            ]
          },
          "limit": {
            "type": "integer",
            "maximum": 50.0,
            "minimum": 1.0,
            "title": "Limit",
            "description": "Maximum number of recommendations to return.",
            "default": 10,
            "examples": [
              8
            ]
          },
          "include_accommodation": {
            "type": "boolean",
            "title": "Include Accommodation",
            "description": "Whether accommodation objects should be included in recommendations.",
            "default": false,
            "examples": [
              false
            ]
          },
          "include_categories": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Include Categories",
            "description": "Optional category allow-list.",
            "examples": [
              [
                "history",
                "culture",
                "food",
                "nature"
              ]
            ]
          },
          "exclude_categories": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Exclude Categories",
            "description": "Optional category deny-list.",
            "examples": [
              [
                "shopping"
              ]
            ]
          }
        },
        "type": "object",
        "required": [
          "trip_request_id"
        ],
        "title": "RecommendationGenerateRequest"
      },
      "RecommendationListResponse": {
        "properties": {
          "recommendation_run_id": {
            "type": "integer",
            "title": "Recommendation Run Id",
            "description": "Persisted recommendation run id.",
            "examples": [
              1
            ]
          },
          "trip_request_id": {
            "type": "integer",
            "title": "Trip Request Id",
            "description": "Trip request id.",
            "examples": [
              1
            ]
          },
          "city_id": {
            "type": "integer",
            "title": "City Id",
            "description": "City id.",
            "examples": [
              6
            ]
          },
          "total_candidates": {
            "type": "integer",
            "title": "Total Candidates",
            "description": "Number of candidate POIs before limiting.",
            "examples": [
              7
            ]
          },
          "recommendations": {
            "items": {
              "$ref": "#/components/schemas/RecommendationResponse"
            },
            "type": "array",
            "title": "Recommendations"
          }
        },
        "type": "object",
        "required": [
          "recommendation_run_id",
          "trip_request_id",
          "city_id",
          "total_candidates",
          "recommendations"
        ],
        "title": "RecommendationListResponse"
      },
...
      "TripLocationInput": {
        "properties": {
          "address": {
            "type": "string",
            "title": "Address",
            "description": "Human-readable location address.",
            "examples": [
              "ул. Большая Садовая, 121"
            ]
          },
          "latitude": {
            "type": "number",
            "title": "Latitude",
            "description": "Location latitude.",
            "examples": [
              47.2269
            ]
          },
          "longitude": {
            "type": "number",
            "title": "Longitude",
            "description": "Location longitude.",
            "examples": [
              39.7139
            ]
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name",
            "description": "Optional location label.",
            "examples": [
              "Отель"
            ]
          },
          "poi_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Poi Id",
            "description": "Optional accommodation POI id.",
            "examples": [
              6012
            ]
          }
        },
        "type": "object",
        "required": [
          "address",
          "latitude",
          "longitude"
        ],
        "title": "TripLocationInput"
      },
...
      "TripRequestCreate": {
        "properties": {
          "city_id": {
            "type": "integer",
            "title": "City Id",
            "description": "City id for trip planning.",
            "examples": [
              6
            ]
          },
          "profile_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Profile Id",
            "description": "Existing saved profile id to reuse for this trip.",
            "examples": [
              1
            ]
          },
          "profile": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/UserProfileCreate"
              },
              {
                "type": "null"
              }
            ],
            "description": "Inline profile to persist and attach to the trip request."
          },
          "days_count": {
            "type": "integer",
            "maximum": 14.0,
            "minimum": 1.0,
            "title": "Days Count",
            "description": "Trip duration in days.",
            "default": 1,
            "examples": [
              1
            ]
          },
          "daily_time_limit_hours": {
            "type": "integer",
            "maximum": 16.0,
            "minimum": 1.0,
            "title": "Daily Time Limit Hours",
            "description": "Available planning time per day.",
            "default": 8,
            "examples": [
              8
            ]
          },
          "selected_interests": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Selected Interests",
            "description": "Optional interests selected for this specific trip. If omitted, profile interests are used.",
            "examples": [
              [
                "food",
                "history",
                "walking"
              ]
            ]
          },
          "constraints": {
            "additionalProperties": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                },
                {
                  "type": "number"
                },
                {
                  "type": "boolean"
                }
              ]
            },
            "type": "object",
            "title": "Constraints",
            "description": "Additional planning constraints for future extensions.",
            "examples": [
              {
                "avoid_expensive": true
              }
            ]
          },
          "start_location": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TripLocationInput"
              },
              {
                "type": "null"
              }
            ],
            "description": "Preferred start point such as a selected hotel."
          },
          "end_location": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TripLocationInput"
              },
              {
                "type": "null"
              }
            ],
            "description": "Preferred end point. If omitted, the route may return to the start location."
          }
        },
        "type": "object",
        "required": [
          "city_id"
        ],
        "title": "TripRequestCreate"
      },
      "TripRequestResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id",
            "description": "Created trip request id.",
            "examples": [
              1
            ]
          },
          "city_id": {
            "type": "integer",
            "title": "City Id",
            "description": "City id for trip planning.",
            "examples": [
              6
            ]
          },
          "profile_id": {
            "type": "integer",
            "title": "Profile Id",
            "description": "Persisted user profile id.",
            "examples": [
              1
            ]
          },
          "profile": {
            "$ref": "#/components/schemas/UserProfileResponse"
          },
          "days_count": {
            "type": "integer",
            "title": "Days Count",
            "description": "Trip duration in days.",
            "examples": [
              2
            ]
          },
          "daily_time_limit_hours": {
            "type": "integer",
            "title": "Daily Time Limit Hours",
            "description": "Available planning time per day.",
            "examples": [
              8
            ]
          },
          "selected_interests": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Selected Interests",
            "description": "Interests selected for this trip."
          },
          "constraints": {
            "additionalProperties": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                },
                {
                  "type": "number"
                },
                {
                  "type": "boolean"
                }
              ]
            },
            "type": "object",
            "title": "Constraints",
            "description": "Trip-level constraints snapshot."
          },
          "start_location": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TripLocationResponse"
              },
              {
                "type": "null"
              }
            ],
            "description": "Trip start point."
          },
          "end_location": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TripLocationResponse"
              },
              {
                "type": "null"
              }
            ],
            "description": "Trip end point."
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At",
            "description": "Creation timestamp."
          }
        },
        "type": "object",
        "required": [
          "id",
          "city_id",
          "profile_id",
          "profile",
          "days_count",
          "daily_time_limit_hours",
          "created_at"
        ],
        "title": "TripRequestResponse"
      },
...
    }
  },
...
}